refactor(wasm): extract wasmCtx methods
