+++ internal/pkg/wasm/quorum.go
@@ -1,3 +1,4 @@
+//go:build js && wasm
 // +build js,wasm
 
 package wasm
@@ -5,12 +6,8 @@ package wasm
 import (
 	"context"
 	"crypto/rand"
-	"fmt"
-	"sync/atomic"
-	"time"
 
 	ethKeystore "github.com/ethereum/go-ethereum/accounts/keystore"
-	"github.com/libp2p/go-libp2p-core/peer"
 	"github.com/rumsystem/quorum/internal/pkg/chain"
 	"github.com/rumsystem/quorum/internal/pkg/nodectx"
 	"github.com/rumsystem/quorum/internal/pkg/options"
@@ -18,6 +15,9 @@ import (
 	"github.com/rumsystem/quorum/internal/pkg/storage"
 	quorumStorage "github.com/rumsystem/quorum/internal/pkg/storage"
 )
+
+/* global, JS should interact with it */
+var wasmCtx *QuorumWasmContext = nil
 
 func StartQuorum(qchan chan struct{}, bootAddrsStr string) {
 	ctx, cancel := context.WithCancel(context.Background())
@@ -49,82 +49,27 @@ func StartQuorum(qchan chan struct{}, bootAddrsStr string) {
 	//nodectx.GetNodeCtx().PublicKey = keys.PubKey
 	//nodectx.GetNodeCtx().PeerId = peerid
 
+	/* quorum has global groupmgr, init it here */
 	groupmgr := chain.InitGroupMgr(dbMgr)
 
-	err = groupmgr.SyncAllGroup()
-	if err != nil {
-		panic(err)
-	}
+	// TODO: construct app db
 
-	// TODO: app db
-
-	wasmCtx := NewQuorumWasmContext(qchan, config, node, ctx, cancel)
+	wasmCtx = NewQuorumWasmContext(qchan, config, node, ctx, cancel)
 
 	/* Bootstrap will connect to all bootstrap nodes in config.
 	since we can not listen in browser, there is no need to anounce */
-	Bootstrap(wasmCtx)
+	wasmCtx.Bootstrap()
 
 	/* TODO: should also try to connect known peers in peerstore which is
 	   not implemented yet */
 
 	/* keep finding peers, and try to connect to them */
-	go startBackgroundWork(wasmCtx)
-}
+	go wasmCtx.StartDiscoverTask()
 
-func Bootstrap(wasmCtx *QuorumWasmContext) {
-	bootstraps := []peer.AddrInfo{}
-	for _, peerAddr := range wasmCtx.Config.BootstrapPeers {
-		peerinfo, _ := peer.AddrInfoFromP2pAddr(peerAddr)
-		bootstraps = append(bootstraps, *peerinfo)
-	}
-
-	connectedPeers := wasmCtx.QNode.AddPeers(wasmCtx.Ctx, bootstraps)
-	println(fmt.Sprintf("Connected to %d peers", connectedPeers))
-}
-
-func startBackgroundWork(wasmCtx *QuorumWasmContext) {
-	/* first job will start after 1 second */
-	go func() {
-		time.Sleep(1 * time.Second)
-		backgroundWork(wasmCtx)
-	}()
-
-	ticker := time.NewTicker(30 * time.Second)
-	for {
-		select {
-		case <-ticker.C:
-			backgroundWork(wasmCtx)
-		case <-wasmCtx.Qchan:
-			ticker.Stop()
-			wasmCtx.Cancel()
-		}
-	}
-}
-
-func backgroundWork(wasmCtx *QuorumWasmContext) {
-	println("Searching for other peers...")
-	peerChan, err := wasmCtx.QNode.RoutingDiscovery.FindPeers(wasmCtx.Ctx, DefaultRendezvousString)
+	/* start syncing all local groups */
+	err = groupmgr.SyncAllGroup()
 	if err != nil {
 		panic(err)
-	}
-
-	var connectCount uint32 = 0
-
-	for peer := range peerChan {
-		curPeer := peer
-		println("Found peer:", curPeer.String())
-		go func() {
-			pctx, cancel := context.WithTimeout(wasmCtx.Ctx, time.Second*10)
-			defer cancel()
-			err := wasmCtx.QNode.Host.Connect(pctx, curPeer)
-			if err != nil {
-				println("Failed to connect peer: ", curPeer.String())
-				cancel()
-			} else {
-				curConnectedCount := atomic.AddUint32(&connectCount, 1)
-				println(fmt.Sprintf("Connected to peer(%d): %s", curConnectedCount, curPeer.String()))
-			}
-		}()
 	}
 }
